\name{r2dRplot}
\alias{r2dRplot}
\alias{rgf.summarize}
\alias{summarize}

\title{
functions to plot several results of 2dRue analisys
}
\description{
incomplete help
}
\usage{
r2dRplot(o, type = "rain", scope = "run", var = "vi", pixel = 1, 
col = c("blue", "green4", "salmon", "gray"))
}

\arguments{
  \item{o}{
a r2dRue object
}
  \item{type}{
valids type are: "rain", "ndvi", "vimax", "box", "density", "assessment1", assessment2", "monitoring", or "pixel"
}
  \item{scope}{
unused in this implementation
}
  \item{var}{
"vi" or "rain"
}
  \item{pixel}{
unused in this implementation
}
  \item{col}{
unused in this implementation
}
}
