\name{r2dRue-package}
\Rdversion{1.1}
\alias{r2dRue-package}
\alias{r2dRue}
\docType{package}
\title{
2dRUE model functions
}
\description{
2dRUE is a methodology to make a diagnostic of land condition in a large territory during a given time period. Was developed under the DeSurvey IP.
}
\details{
\tabular{ll}{
Package: \tab r2dRue\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2009-12-17\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Gabriel del Barrio, Juan Puigdefabregas, Maria E. Sanjuan & Alberto Ruiz\cr
Maintainer:
Alberto Ruiz, <aruiz@eeza.csic.es>
}
\references{
Article submitted and under revision.
}
